Updated vector resolution code
The vector resolution code was changed to save a new .csv file, named output_data.csv, in the same location as the code. This file will have the resolved accelerations. 

Note that the difference in the names between the two - and possibly any more that may come about later - .csv files could cause issues with the programs as they are saved to here; but that can be amended through simply changing the file used.

(I have also realised that when I made this code I had called it resultant vector calculations, but that is technically wrong so I have corrected it to be vector resolution. This should be in line with what my other programs have it as, but this might cause issues if I have overlooked anything.)